test: Make log tests Windows compatible
Two issues:
1) The log files have to be closed before the temp directory can be deleted.
2) Some tests build upon the Linux permission system. Can't do that in Windows
